transfer restriction to foundations
We keep the donations of the donors in the contract, not in the foundation metamask, and when spending in foundations, they can spend the money allocated to them in the contract. In this way, we oblige foundations to spend through our system, preventing the inability to track donations if they are spent outside the system by foundations. In this way, users can track where their money is spent in any case.